// frontend-test/src/types/app.ts
export type FilterValues = {
    slug: string;

    pageSize: string;
    page: string;

    housing: string[];
    roomsCount: string[];

    priceMin: string;
    priceMax: string;

    squareMin: string;
    squareMax: string;

    floorMin: string;
    floorMax: string;

    roomsMin: string;
    roomsMax: string;
};